Move storage into `retrospective`
