+++ src/flowrep/api/tools.py
@@ -13,5 +13,5 @@ from flowrep.parsers.atomic_parser import parse_atomic as parse_atomic
 from flowrep.parsers.workflow_parser import parse_workflow as parse_workflow
 from flowrep.parsers.workflow_parser import workflow as workflow
 from flowrep.retrospective.datastructures import recipe2data as recipe2data
-from flowrep.storage import LexicalBagBrowser as LexicalBagBrowser
+from flowrep.retrospective.storage import LexicalBagBrowser as LexicalBagBrowser
 from flowrep.wfms import run_recipe as run_recipe
