+++ src/flowrep/retrospective/storage.py
@@ -12,8 +12,8 @@ from typing import TYPE_CHECKING
 from packaging import version
 from pyiron_snippets import import_alarm
 
-from flowrep import base_models, storage_widget
-from flowrep.retrospective import datastructures
+from flowrep import base_models
+from flowrep.retrospective import datastructures, storage_widget
 
 with import_alarm.ImportAlarm(
     "This tool requires the 'bagofholding' package.", raise_exception=True
